Move common graphics utils to //flutter/common/graphics
This change creates a new source set //flutter/common/graphics
with minimal dependencies.

The intention with this target is to be later consumed in a
darwin common graphics target across the desktop boundary.

Also unifies some of the graphics utilities under the same folder.
