+++ common/graphics/persistent_cache.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "flutter/shell/common/persistent_cache.h"
+#include "flutter/common/graphics/persistent_cache.h"
 
 #include <future>
 #include <memory>
