+++ common/graphics/persistent_cache.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef FLUTTER_SHELL_COMMON_PERSISTENT_CACHE_H_
-#define FLUTTER_SHELL_COMMON_PERSISTENT_CACHE_H_
+#ifndef FLUTTER_COMMON_GRAPHICS_PERSISTENT_CACHE_H_
+#define FLUTTER_COMMON_GRAPHICS_PERSISTENT_CACHE_H_
 
 #include <memory>
 #include <mutex>
@@ -134,4 +134,4 @@ class PersistentCache : public GrContextOptions::PersistentCache {
 
 }  // namespace flutter
 
-#endif  // FLUTTER_SHELL_COMMON_PERSISTENT_CACHE_H_
+#endif  // FLUTTER_COMMON_GRAPHICS_PERSISTENT_CACHE_H_
