+++ common/graphics/texture.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef FLUTTER_FLOW_TEXTURE_H_
-#define FLUTTER_FLOW_TEXTURE_H_
+#ifndef FLUTTER_COMMON_GRAPHICS_TEXTURE_H_
+#define FLUTTER_COMMON_GRAPHICS_TEXTURE_H_
 
 #include <map>
 
@@ -74,4 +74,4 @@ class TextureRegistry {
 
 }  // namespace flutter
 
-#endif  // FLUTTER_FLOW_TEXTURE_H_
+#endif  // FLUTTER_COMMON_GRAPHICS_TEXTURE_H_
