+++ shell/common/BUILD.gn
@@ -69,8 +69,6 @@ source_set("common") {
     "display_manager.h",
     "engine.cc",
     "engine.h",
-    "persistent_cache.cc",
-    "persistent_cache.h",
     "pipeline.cc",
     "pipeline.h",
     "platform_view.cc",
@@ -111,6 +109,7 @@ source_set("common") {
   deps = [
     "//flutter/assets",
     "//flutter/common",
+    "//flutter/common/graphics",
     "//flutter/flow",
     "//flutter/fml",
     "//flutter/lib/ui",
@@ -189,6 +188,7 @@ if (enable_unittests) {
     ]
 
     public_deps = [
+      "//flutter/common/graphics",
       "//flutter/flow",
       "//flutter/fml/dart",
       "//flutter/runtime",
@@ -253,6 +253,7 @@ if (enable_unittests) {
       ":shell_test_fixture_sources",
       ":shell_unittests_fixtures",
       "//flutter/assets",
+      "//flutter/common/graphics",
       "//flutter/shell/version",
       "//third_party/googletest:gmock",
     ]
