+++ shell/common/shell_test_platform_view_vulkan.cc
@@ -4,7 +4,7 @@
 
 #include "flutter/shell/common/shell_test_platform_view_vulkan.h"
 
-#include "flutter/shell/common/persistent_cache.h"
+#include "flutter/common/graphics/persistent_cache.h"
 #include "flutter/vulkan/vulkan_utilities.h"
 
 namespace flutter {
