+++ shell/platform/android/android_external_texture_gl.h
@@ -7,7 +7,7 @@
 
 #include <GLES/gl.h>
 
-#include "flutter/flow/texture.h"
+#include "flutter/common/graphics/texture.h"
 #include "flutter/fml/platform/android/jni_weak_ref.h"
 #include "flutter/shell/platform/android/platform_view_android_jni_impl.h"
