+++ shell/platform/darwin/ios/ios_external_texture_metal.h
@@ -9,7 +9,7 @@
 
 #import <CoreVideo/CoreVideo.h>
 
-#include "flutter/flow/texture.h"
+#include "flutter/common/graphics/texture.h"
 #include "flutter/fml/macros.h"
 #include "flutter/fml/platform/darwin/cf_utils.h"
 #include "flutter/fml/platform/darwin/scoped_nsobject.h"
